requirements.txt
sympy
pwntools
Pillow
PyJWT
pycryptodome
scapy
aiohttp